requirements-dev.txt
pre-commit
pytest
pytest-mock
black
SQLAlchemy==1.4.27
passlib==1.7.4
python-jose[cryptography]==3.3.0